+++ testapp/testapp/testmain/tests/test_views.py
@@ -171,7 +171,6 @@ class ReceiptPDFTestCase(TestCase):
         """
         taxpayer = factories.TaxPayerFactory()
 
-        factories.TaxPayerProfileFactory(taxpayer=taxpayer)
         pdf = factories.ReceiptPDFFactory(
             receipt__point_of_sales__owner=taxpayer,
         )
@@ -190,7 +189,6 @@ class ReceiptPDFTestCase(TestCase):
 @pytest.mark.django_db
 def test_template_discovery(client):
     taxpayer = factories.TaxPayerFactory(cuit="20329642330")
-    factories.TaxPayerProfileFactory(taxpayer=taxpayer)
     pdf = factories.ReceiptPDFFactory(
         receipt__point_of_sales__owner=taxpayer,
         receipt__point_of_sales__number=9999,